Safari-611.1.10/LayoutTests/media/picture-in-picture/picture-in-picture-api-enter-pip-2-expected.txt

This tests that request Picture-in-Picture requires loaded metadata for the video element.

RUN(internals.settings.setAllowsPictureInPictureMediaPlayback(true))
RUN(internals.setMockVideoPresentationModeEnabled(true))
EXPECTED (error.name == 'InvalidStateError') OK
END OF TEST
